requirements.txt
asgiref==3.7.2
Django==4.1.13
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.0
iniconfig==2.0.0
packaging==23.2
Pillow==9.2.0
pluggy==1.4.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==8.0.2
pytest-django==4.8.0
python-decouple==3.8
pytz==2024.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
